mom/requirements.txt
beautifulsoup4==4.12.2
click==8.1.3
Flask==2.2.3
google==3.0.0
grpcio==1.54.0
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
protobuf==4.22.3
soupsieve==2.4.1
Werkzeug==2.2.3
zipp==3.15.0
